refactor: moves component to it's own directory
Relates #36
